Refactor class abstraction
