+++ service/job/test-steps/src/main/java/org/eclipse/kapua/service/job/steps/model/TestJobStepProcessor.java
@@ -15,6 +15,7 @@ package org.eclipse.kapua.service.job.steps.model;
 
 import org.eclipse.kapua.KapuaException;
 import org.eclipse.kapua.commons.security.KapuaSecurityUtils;
+import org.eclipse.kapua.job.engine.commons.operation.AbstractDeviceTargetProcessor;
 import org.eclipse.kapua.job.engine.commons.operation.AbstractTargetProcessor;
 import org.eclipse.kapua.job.engine.commons.wrappers.JobTargetWrapper;
 import org.eclipse.kapua.locator.KapuaLocator;
@@ -27,9 +28,7 @@ import javax.batch.runtime.context.JobContext;
 import javax.batch.runtime.context.StepContext;
 import javax.inject.Inject;
 
-public class TestJobStepProcessor extends AbstractTargetProcessor implements TargetProcessor {
-    private static final KapuaLocator LOCATOR = KapuaLocator.getInstance();
-    private static final DeviceRegistryService DEVICE_REGISTRY_SERVICE = LOCATOR.getService(DeviceRegistryService.class);
+public class TestJobStepProcessor extends AbstractDeviceTargetProcessor implements TargetProcessor {
     @Inject
     JobContext jobContext;
 
@@ -49,14 +48,5 @@ public class TestJobStepProcessor extends AbstractTargetProcessor implements Tar
         if (fail) {
             throw KapuaException.internalError("This processing has been set to fail");
         }
-    }
-
-    @Override
-    protected String getTargetDisplayName(JobTarget jobTarget) throws KapuaException {
-        Device device = KapuaSecurityUtils.doPrivileged(() -> DEVICE_REGISTRY_SERVICE.find(jobTarget.getScopeId(), jobTarget.getJobTargetId()));;
-        if (device == null) {
-            return "N/A";
-        }
-        return device.getClientId();
     }
 }
